Fixing based on Jim’s suggestions:
Update button in Contact section to Message the Nunemaker Lab

On the first Projects page, I feel like there should be something above 'what are the islets for langerhans' Maybe an "Our Focus:" headline?

For the 3 individual project pages, I'd like to see a back to projects link at the top too.

On the inside pages, Craig's email isn't showing with the envelope icon.